# go_sprint/api/openapi/combined-with-auth.yaml
openapi: 3.0.3
info:
  title: Sprint Management API with Authentication
  version: 1.0.0
  description: Complete API for sprint management including authentication
servers:
- url: http://localhost:8082
  description: Local development server
paths:
...
  /api/v1/sprints:
    get:
      summary: List sprints
      description: Retrieve a paginated list of sprints with optional filtering
      operationId: listSprints
      tags:
      - Sprints
      parameters:
      - name: status
        in: query
        description: Filter by sprint status
        schema:
          $ref: '#/components/schemas/SprintStatus'
      - name: start_date_after
        in: query
        description: Filter sprints starting after this date
        schema:
          type: string
          format: date
      - name: start_date_before
        in: query
        description: Filter sprints starting before this date
        schema:
          type: string
          format: date
      - name: cursor
        in: query
        description: Cursor for pagination
        schema:
          type: string
      - name: limit
        in: query
        description: Number of items per page (max 100)
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create sprint
      description: Create a new sprint with date range and capacity
      operationId: createSprint
      tags:
      - Sprints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSprintRequest'
      responses:
        '201':
          description: Sprint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}:
    get:
      summary: Get sprint
      description: Retrieve a specific sprint by ID with all details
      operationId: getSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update sprint
      description: Update an existing sprint
      operationId: updateSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSprintRequest'
      responses:
        '200':
          description: Sprint updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sprint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete sprint
      description: Soft delete a sprint
      operationId: deleteSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Sprint deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}/close:
    post:
      summary: Close sprint
      description: Close a sprint and move incomplete work items to backlog
      operationId: closeSprint
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Sprint closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintCloseResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}/workitems:
    get:
      summary: Get sprint work items
      description: Retrieve all work items assigned to a sprint
      operationId: getSprintWorkItems
      tags:
      - Sprints
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        description: Filter by work item status
        schema:
          type: string
          enum:
          - todo
          - in_progress
          - in_review
          - done
          - blocked
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/sprints/{id}/burndown:
    get:
      summary: Get sprint burndown chart data
      description: Retrieve burndown chart data for a sprint
      operationId: getSprintBurndown
      tags:
      - Sprints
      - Metrics
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BurndownData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/{id}/velocity:
    get:
      summary: Get sprint velocity
      description: Calculate velocity for a completed sprint
      operationId: getSprintVelocity
      tags:
      - Sprints
      - Metrics
      parameters:
      - name: id
        in: path
        required: true
        description: Sprint ID
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VelocityData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/planning/capacity:
    post:
      summary: Calculate sprint capacity
      description: Calculate recommended sprint capacity based on team velocity
      operationId: calculateSprintCapacity
      tags:
      - Sprints
      - Planning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapacityCalculationRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityRecommendation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/sprints/planning/forecast:
    post:
      summary: Forecast sprint completion
      description: Forecast sprint completion based on historical velocity
      operationId: forecastSprintCompletion
      tags:
      - Sprints
      - Planning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForecastRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    SprintStatus:
      type: string
      enum:
      - planned
      - active
      - completed
      - cancelled
      description: Current status of sprint
    Sprint:
      type: object
      required:
      - id
      - name
      - status
      - start_date
      - end_date
      - created_by
      - created_at
      - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
        name:
          type: string
          maxLength: 255
          description: Sprint name
        description:
          type: string
          description: Sprint description
        status:
          $ref: '#/components/schemas/SprintStatus'
        start_date:
          type: string
          format: date
          description: Sprint start date
        end_date:
          type: string
          format: date
          description: Sprint end date
        capacity_points:
          type: integer
          minimum: 0
          nullable: true
          description: Sprint capacity in story points
        committed_points:
          type: integer
          minimum: 0
          default: 0
          description: Committed story points
        completed_points:
          type: integer
          minimum: 0
          default: 0
          description: Completed story points
        created_by:
          type: string
          format: uuid
          description: User who created the sprint
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
    SprintDetail:
      allOf:
      - $ref: '#/components/schemas/Sprint'
      - type: object
        properties:
          work_items:
            type: array
            items:
              $ref: '#/components/schemas/WorkItemSummary'
            description: Work items in this sprint
          metrics:
            $ref: '#/components/schemas/SprintMetrics'
            description: Sprint metrics
    CreateSprintRequest:
      type: object
      required:
      - name
      - start_date
      - end_date
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Sprint name
        description:
          type: string
          description: Sprint description
        start_date:
          type: string
          format: date
          description: Sprint start date
        end_date:
          type: string
          format: date
          description: Sprint end date (must be after start_date)
        capacity_points:
          type: integer
          minimum: 0
          nullable: true
          description: Sprint capacity in story points
    UpdateSprintRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Sprint name
        description:
          type: string
          description: Sprint description
        start_date:
          type: string
          format: date
          description: Sprint start date
        end_date:
          type: string
          format: date
          description: Sprint end date (must be after start_date)
        capacity_points:
          type: integer
          minimum: 0
          nullable: true
          description: Sprint capacity in story points
        status:
          $ref: '#/components/schemas/SprintStatus'
...
    SprintListResponse:
      type: object
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Sprint'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict - business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- BearerAuth: []